Account for null values in creators
